Auth-Interface Login change
Now waits for the 'GetAuthSessionTicketResponse_t' callback, which will be called when this ticket has been communicated to the server (Steam I believe).

The ticket is ready to use after this callback, so the LoginAuth method has been changed to wait for this callback and only login when the ticket is ready to be used.