Local/Online user states. And user data from newly joined users.
Users receive an event when someone does something in the lobby.
Users joining the lobby will cause the users within the lobby to load some info about them, like username, avatar, and external accounts info. Also has a failsafe and checks if the user is still in the lobby after retrieving this data.